Testes com Heroku

+++ src/main/java/hello/User.java
@@ -9,17 +9,17 @@ import javax.persistence.Id;
 public class User {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     private String nome;
 
     private String telefone;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
@@ -39,7 +39,6 @@ public class User {
 		this.telefone = telefone;
 	}
 
-	
+
 
 }
-
